feat: [MR-549] Encode message kind and context in `MessageId`
Encode message kind (request or response) in the `MessageId` instead of relying
on a separate `MessagePoolReference`. Also encode the context (inbound or
outbound) in the `MessageId`; this makes it easier (and more efficient) to
update the stats when a message is timed out or shed.

Also move part of `MemoryUsageStats` (guaranteed response sizes, guaranteed
requests extra bytes, best-effort message sizes) from `CanisterQueues` into
`MessagePool`. `CanisterQueues` will continue to track guaranteed response
reservations and the transient responses in stream size. (The `CanisterQueues`
change will be coming separately.)