feat: [MR-550] `MessagePool` persistence
Implement `MessagePool` encoding and decoding.

As a side-effect, also change the types of the two priority queues from
`BinaryHeap` to `BTreeSet`: insertion and removal from the pool was already
`O(log(N))`, so we may as well simplify the logic and ensure that the priority
queues are always in sync with the contents of the pool This requires explicitly
recording (and persisting) the implicitly assigned deadlines of outbound
guaranteed response requests. But it also means that the priority queues do not
need to be explicitly persisted.