feat: [IC-1754] Rename `CanisterQueues` stats to reflect slot vs memory reservations
Since we're moving into a world where queue slot reservations (used by the
backpressure logic) are different from queue response memory reservations
(ensuring guaranteed response delivery), rename the relevant  stats methods to
more accurately reflect the kind of reservation they track.

This is in preparation for the `CanisterQueues` rewrite, in order to reduce the
size of that diff.