chore(IDX): introduce the push_to_gitlab_improved job
The `push_to_gitlab_improved` GitHub job is an improved version of
`push_to_gitlab` which creates a MR for just the HEAD of the
`mirroring` branch. `push_to_gitlab` actually creates a MR for the
whole `mirroring` branch which sometimes contains commits of multiple
PRs.

Once `push_to_gitlab_improved` is working as expected we'll remove the old
`push_to_gitlab` and rename `push_to_gitlab_improved` to it.