chore(IDX): remove bazel-build-log.json
This changes the default config to not general a build log. The build
log is a very large file, and is not used much.

When debugging builds, the build log can simply be re-enabled. It does
not seem to be generally useful to generate the build log or to store
it.