fix(IDX): check PR title compliance always
This changes the conventional commit check to use the default
(suggested) config. This means the behavior is consistent regardless of
whether the PR is a draft or not. Before this, an edit (base change,
description edit) of a draft PR would trigger a run of the action. Now,
the action is always run on draft and non-draft PRs.

This also adds the action's version next to its commit sha for
visibility, and removes the concurrency settings (as the action is quick
to run and not resource-hungry and to match the action's recommended
settings).