chore: CRP-2509 add `additional_colocate_flag` and fetch targets via `bazel query` to fix `system-tests-benchmarks-nightly` (#263)
The relevant, currently only one, test in the benchmark pipeline was not
invoked because it is tagged as `manual`, so `test_tag_filter` won't
find it. This PR changes the logic to use `bazel query` that finds all
tagged targets and run `bazel test` on its outputs directly. Also, this
PR adds `additional_colocate_tags` to the `system_test` bazel function
to pass the tag only to the colocated test and not both
`tecdsa_performance` and `tecdsa_performance_colocate`.

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>
Co-authored-by: Marko Kosmerl <marko.kosmerl@dfinity.org>