fix: RUN-1014: Free SandboxedExecutionController threads (#354)
This fixes two thread leaks per SandboxedExecutionController. Note that
in production replica that fix has little impact because there is only
one controller that has the same lifetime as the process itself.

However, in tests that create many instances of the controller, this
considerably reduces the number of threads.

Kudos to mraszyk@ for reporting the issue and providing a way to
reproduce the leak and verify the fix.

Fixes:

- the sandbox monitoring thread was not terminated properly. The fix
introduces a channel to request termination when the controller is
dropped.

- the background writing thread of the launcher service was not
terminated properly. The fix terminates the writing thread when the
reading thread exits (similar to other places).