chore: firewall counter exporter (#343)
The replica node and API boundary node firewall has connection- and
rate-limits. At the moment, each packet that is dropped based on these
rules is logged.

In this PR, we switch to counters and add a service to export the
counters as a metric in order to not spam logs and to easily monitor
whether the rules are actually active.

---------

Co-authored-by: IDX GitLab Automation <infra+gitlab-automation@dfinity.org>