test(icp_ledger): FI-1301, FI-1359: Get and query all blocks from ledger and archives and fix test_archive_indexing (#398)
Get and query all the blocks from the ICP ledger and archives in the
`icp_get_blocks` and `icp_query_blocks` helper functions: Try to
retrieve all the blocks from the ledger, and if there are archives,
retrieve all the blocks from them also, up to the maximum blocks per
response at a time. Previously, only one call to retrieve blocks to the
ledger and archives was made, so if any of them contained more blocks
than could fit into a response, not all were retrieved.

Also fix the ICP `test_archive_indexing` test that broke.