perf: CON-1360 Reduce cost of cloning tSchnorr inputs (#344)
The threshold Schnorr request context contains the message to be signed,
which may be up to 10Mib in size. When iterating over open request
contexts and building the signature inputs, this message is currently
being cloned
[here](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/consensus/src/ecdsa/utils.rs?L292).

Although preliminary performance tests with a large number of maximum
size tSchnorr request didn’t reveal any degradation in finalization
rate, wrapping this message in an `Arc` reduced the average
`on_state_change` duration by a factor of 2.