chore(PocketIC): refactor progress threads in PocketIC (#353)
This MR refactors progress threads in PocketIC by introducing a helper
function `execute_operation` that drives an operation on a PocketIC
instance to completion. Such a helper function will be useful for
introducing canister HTTP outcalls in a follow up MR where the progress
threads will also need to execute operations related to canister HTTP
outcalls and thus factoring out the logic of driving an operation to
completion into a separate helper function is beneficial.