chore(nns): Add api <--> internal type conversions (#393)
### Why
To improve build times, we want other parts of the build to not depend
on the entire governance crate. Additionally, we want to separate API
types from internal governance types so that they can evolve
independently. This is a first mechanical step towards that end.
### Changes
This MR creates type conversions from governance types to api types. It
does not yet use them, but allows Governance to start using them (in
followup MRs) in the main canister file.