feat(ckerc20): Use EVM-RPC canister to call `eth_getLogs` (#400)
([XC-144](https://dfinity.atlassian.net/browse/XC-144)): Continue the
migration initiated in
[MR-19972](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/19972)
towards using the EVM-RPC canister to interact with the Ethereum
JSON-RPC providers. This PR focuses specifically on the `eth_getLogs`
endpoint, which when queried by the EVM-RPC canister brings the
following changes in comparison to when the call is made directly by the
minter:
1. The maximum range of blocks whose logs can be retrieved is lower
(`500` instead of `799`)
2. The queried topics have an additional level of nesting
(e.g.,`[["0x257e..."]]` is queried instead of `["0x257e..."]`)
3. The `max_response_bytes` value given to make HTTPs outcalls is
adjusted differently in case the response is bigger than expected (e.g.
`2 * a` instead of `2 * b + HEADER_SIZE_LIMIT`, where `a`, `b` denote
the previous estimates)

[XC-144]:
https://dfinity.atlassian.net/browse/XC-144?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ