fix: Maintain snapshot_ids correctly (#360)
This commit fixes an issue with how `CanisterSnapshots` maintains
`snapshot_ids`, which is a map from canisters to the set of snapshots it
owns.

The constructor now computes `snapshot_ids` from `snapshots`, instead of
taking them as an argument. This ensures that the two fields do not
contradict each other.