chore(IDX): remove targets from rust_canister (#440)
This reduces the number of targets that `rust_canister` generates. In
particular:

* The bin (transition base) `rust_binary` build is marked as `"manual"`
so that the native build is not performed when e.g. running `bazel build
//rs/...`
* A lot of intermediary targets (gzipping, metadata, shrinking) are
removed and replaced with a single `finalize_wasm` pass, which helps
decrease the total number of targets (less data duplicated, fewer cache
downloads, etc)

This saves ~300 targets compared to previously (2305 -> 1991), including
strictly unnecessary native builds of canisters.

The transition `whitelist` attribute is also replaced with `allowlist`
(the former has been deprecated for a while).

---------

Co-authored-by: Bas van Dijk <bas@dfinity.org>