feat(nns): Known neurons are always public. (#488)
This is done by adding a normalization step towards the end of fetching
a neuron from NeuronStore. More precisely, normalized(neuron) returns a
possibly modified copy of neuron where the visibility field is set to
Some(Visibility::Public) when the known_neuron_data field is Some(...).

This approach requires no data migration (and no extra storage). One
drawback though is that when visibility = null in a stored neuron, it
means something different depending on whether the neuron is known or
not: known + visibility=null means public; whereas not known +
visibility=null means private. This seems ok, because in our design, we
intuitively have two rules:

1. By default, neurons are private.
2. Known neurons are always public.

Closes [NNS1-3074](https://dfinity.atlassian.net/browse/NNS1-3074).

[<< Previous PR](https://github.com/dfinity/ic/pull/451)

[NNS1-3074]:
https://dfinity.atlassian.net/browse/NNS1-3074?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ