refactor(IDX): simplify hotfix branch matching logic (#514)
This PR slightly changes the procedure of security hotfixes, where
branches now only only need to follow the pattern `hotfix-*` instead of
`hotfix-.*-rc--.*`. The reasons for this change are:

a) We have run into a lot of confusion in the past because of incorrect
branch names
b) It lets us unify the hotfix process for both regular and ic-os (?)
hotfixes

This change is backwards compatible, as branches with the
`hotfix-.*-rc--.* ` will still behave the same.