refactor(nervous_system): Add `controller` and `hotkeys` fields to CfParticipant, CfNeuron, and CfInvestment (#373)
## Change 1: 

A `controller` field has been added to replace `hotkey_principal` (which
is now deprecated and will hopefully be removed). New versions of the
code populate both fields. New versions of the code only require at
least one of the two fields to be populated.

This change is useful because the previous name for `controller`,
`hotkey_principal`, is very confusing because it doesn't refer to the
NNS neuron's hotkey, instead it refers to the NNS neuron's controller.
(It is named that way because it becomes a hotkey on the SNS neuron that
is allocated to that NNS neuron, but it does not represent a hotkey on
the NNS neuron.)

## Change 2: 

A `hotkeys` field has been added.

This is useful because we want to transfer NF neurons' hotkeys from the
NNS over to the SNS.

## Upgrade dependency

**None.** This MR merely extends various API with additional fields, and
doesn’t change existing fields. All the new fields are optional, so
there is no upgrade dependency introduced by this MR. No canisters
currently require these fields to be `Some`.

NNS1-3198 (which tracks the removal `hotkey_principal`) can only be
implemented once the old fields are no longer used by any canisters, and
will have to be rolled out carefully since the field is non-optional.

## Misc.

1. I moved the proto definition of `Principals` to nervous_system/proto.
I think this is backwards compatible but the buf backwards compatibility
checker disagrees. Since we haven't deployed anything that uses
Principals yet, it should still be safe to merge