feat(nns): Support StopOrStartCanister proposal action (#458)
# Why

The StopOrStartCansiter NnsFunction was within the
NetworkCanisterManagement topic, and we'd like to split it into
different topics based on the kind of canister being started/stopped. We
now support it as a proposal action instead of an NnsFunction.

# Alternatives
The StopOrStartCansiter proposals are very rare, so it might not seem
like it's worth changing it into a proposal action. However,
alternatives are worse:

1. Stop supporting `StopOrStartCansiter` - even though it's rare, it
doesn't seem justified to remove the support
2. Do not assign different topics for stopping or starting different
canisters - we'd like to make distinctions between "protocol canisters"
and others.
3. Keep supporting it in the form of ExecuteNnsFunction - in order to
know the canister id being stopped or started, we'd have to parse the
opaque payload `StopOrStartCanisterRequest`, which makes NNS Governance
depend on the type (used in NNS Root) anyway. Also, the
`StopOrStartCansiter` NnsFunction can no longer determine its topic,
which makes the `topic()` function even more complicated (look at action
-> nns function -> canister id from opaque blob).


# Tests
An integration test in root.rs is rewritten as a state machine test, and
both cases (NnsFunction and new proposal action) are tested.