chore(management-canister): Remove unimplemented delete_chunks API (#537)
This API is not included in the public interface spec and there are no
plans to implement it in the foreseeable future (at least not until
feedback is received that it's needed). There's already a
`clear_chunk_store` API which should allow clearing the store in an all
or nothing approach, if there's need for finer grain control, the
`delete_chunks` API can be introduced and implemented then.