chore: Remove unused instruction limits from subnet record (#441)
The instructions limits in the subnet record were added a long time ago
but were never used (the limits are set in the replica code). This PR
removes them.

The change is safe to do as far as protobuf is concerned. Given that the
limits appear also in the `CreateSubnetPayload` type which is a
`CandidType`, with the fields being non-optional, it's technically a
breaking change to remove them. The only client using this payload is
`ic-admin` which will maintain a copy of the type including the fields
until the Registry canister is upgraded and does not expect them
anymore.