refactor: Remove misleading `callback_id` from `register_callback()` test function (#497)
The function `register_callback()` takes a `CallbackId` as an argument.
This suggests that this `CallbackId` is registered in the
`CallContextManager` such that a response using this `CallbackId` can
later be inducted successfully by passing this [sanity
check](https://sourcegraph.com/github.com/dfinity/ic@rumenov/regtypes/-/blob/rs/replicated_state/src/canister_state/system_state/call_context_manager.rs?L599).

This is false. The `CallbackId` handed to `register_callback()` is used
to construct an update call, but the origin of a call is irrelevant for
the [sanity
check](https://sourcegraph.com/github.com/dfinity/ic@rumenov/regtypes/-/blob/rs/replicated_state/src/canister_state/system_state/call_context_manager.rs?L599)
because each message gets its own `CallbackId`. It is in general not
possible to ask the `CallContextManager` to register a `CallbackId`,
rather you must ask the `CallContextManager` to register a new
`Callback` for your call and then the `CallContextManager` generates and
registers this `Callback` with a new `CallbackId` and returns it to you.
This is the `CallbackId` that must be used to generate the message to
pass the [sanity
check](https://sourcegraph.com/github.com/dfinity/ic@rumenov/regtypes/-/blob/rs/replicated_state/src/canister_state/system_state/call_context_manager.rs?L599).

Since the origin of the call is irrelevant for the intended purpose of
this function, it makes sense to use a system task as the call origin
(i.e. a heartbeat or a timer), since that does not require a
`CallbackId` as an input and return the `CallbackId` that must be used
to construct the message instead.

The changes in this PR are such that the misleading `CallbackId`
argument is removed, but otherwise the code is intact. Further changes
to improve consistency with how the `CallContextManager` should be used
will be done in a different PR.