refactor(nns): no longer generate api types from internal protos (#588)
We separate the API types used in canister.rs from the internal types
used in nns/governance. This helps both with speeding up the builds
(because we can remove dependencies on nns/governance) and with allowing
independent evolution of the internal domain types and the API, which
will allow us to make better APIs while still supporting older versions
of the APIs.