refactor: introduce artifact downloader component (#403)
Adds trait that allows transforming the artifact before sending it on
the wire and reconstructing the original artifact on the receive side.
The benefit of this is that it allows for artifact specific
transform/reconstruction logic. This will be used for the blocks in
hashes optimization to add block reconstruction.

There should be no changes to any production code in this MR besides the
refactoring.

---------

Co-authored-by: Daniel Sharifi <40335219+DSharifi@users.noreply.github.com>