test(orchestrator): increase subnet size for sr_app_large_with_tecdsa_test to 37 (#586)
Increases the subnet size used in the large app subnet recovery test
with tECDSA from 34 to 37.

One purpose of this test is to showcase how large app subnets with (all)
chain keys enabled can be so that they can still be recovered. In local
tests, the test succeeded around 50 times for 37 nodes, which seems
sufficient for increasing the value permanently.