fix(consensus): [CON-1389] Fix inconsistency when purging validated pool below maximum element (#598)
This MR fixes an edge case in `tx_purge_index_below`. If we purge the
LMDB pool below the maximum element of the given artifact type, the
metadata DB entry for that artifact type would be prematurely deleted
(while the artifact db correctly retains the element). Replacing `<=`
with `<` fixes this problem.