feat(IDX): team labels (#558)
Add `@team` labels (~tags) on PR when and only when PR is ready and when
new commits were added in the PR that added additional teams. We retain
the team tag in the case where the part of the codebase gets removed and
doesn't require a specific team's approval anymore (the same way as
GitHub doesn't remove a team under the Reviewers list unless someone
from that team already engaged with the PR). Having this logic would
would require much more time so it's better to delay it and see if
experience with this is good enough.