feat(nns): Disallow SetVisibility ManageNeuron proposals. (#643)
Because this is new, and clients (e.g. nns-dapp) do not yet have the
latest governance.did to fully deserialize these.

(Technically, clients can partially deserialize such proposals. Also,
technically, clients could dynamically fetch the latest governance.did
file, but this is a more advanced technique and not the typical way to
do things.).

Closes [NNS1-3224](https://dfinity.atlassian.net/browse/NNS1-3224).

This will be reverted per
[NNS1-3228](https://dfinity.atlassian.net/browse/NNS1-3228).

[NNS1-3228]:
https://dfinity.atlassian.net/browse/NNS1-3228?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[NNS1-3224]:
https://dfinity.atlassian.net/browse/NNS1-3224?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ