refactor(sns): Remove the open method from swap. [override-didc-check] (#454)
The only principal that was allowed to call this method was NNS
governance, but it no longer does that. It used to be the main thing
that the OpenSnsTokenSwap proposal called, but that type of proposal can
no longer be used.

Ofc, this is technically a backwards incompatible change, because we
removed a method. However, it is safe, because previously, there was
only one allowed caller, and that caller has stopped calling the removed
method.