refactor(sns): Add controller and hotkeys information to ClaimSwapNeuronsRequest, and use it in SNS Governance (#596)
Note that we transition to an entirely new type inside
`ClaimSwapNeuronsRequest`, `NeuronRecipes` rather than
`NeuronParameters`. However the types are generally very similar
isomorphic.

## Change 1:
A `oneof` has been added to collect information that is unique to direct
participants and neurons' fund participants.

## Change 2:
They `hotkey` field has been replaced with `nns_neuron_controller`. This
is convenient because otherwise the next change would have been
confusing. The old field was named that way because it becomes a hotkey
on the SNS neuron that is allocated to that NNS neuron, but it does not
represent a hotkey on the NNS neuron.

## Change 2:
A hotkeys field has been added.

This is useful because we want to transfer NF neurons' hotkeys from the
NNS over to the SNS.

## Change 3
SNS Governance has been updated to support the new type. The old type is
still supported as of this MR. SNS Swap still sends uses the old type.

## Upgrade dependency
None

NNS1-3198 (which tracks the removal NeuronParameters) can only be
implemented once the old field, `neuron_parameters`, is no longer used
by any canisters, and will have to be rolled out carefully since the
`neuron_parameters` field is non-optional.

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>