feat: Reduce DTS slice limit for regular messages on system subnets (#621)
This is a follow-up change to further decrease the DTS slice limit on
system subnets for regular messages. The value is set to 2B instructions
to match the setting on application subnets.

`install_code` messages still use the bigger slice limit until the
checkpoint interval is adjusted and we can afford reducing that as well.
This will change in a future PR.