chore(backup): [CON-1393] Check if the disk usage exceeds threshold only after running ic-replay (#680)
Currently we check the disk usage every 5 minutes *and* every time we
run ic-replay, and report a warning to slack when the remaining disk
space is below a threshold.

With this change we will only send such a notification after we run
ic-replay.

Also added a "storage type" label to the disk usage metrics so we know
whether the metric is about the cold or hot storage (without having to
look at the directory name)