chore(IDX): create minimal image (#682)
This PR introduces a minimal runner images that can be used across all
dfinity orgs. The reason for this is that self-hosted runners must use a
container image, but using `build-ic` for all jobs not in the ic repo is
impractical. Because the image is public it can be used in other orgs as
well.