fix(nns): Fixed a bug where known neuron is not seen as public. (#699)
The problem is that there are (many) paths where normalization is not
applied. This is a bug from a recent change that I did.

The solution (implemented here) is to make the Neuron::visibility field
private, and add a (public) getter.

This obsoletes the previous solution: `fn normalized`. Thus, I deleted
`normalized`.

This bug was discovered by strengthening an existing test (for public
neuron metrics), which then started failing.