test: Rename COW safety test (#678)
This commit renames the cow_safety_test to memory_safety_test.

COW (copy-on-write) was an old experimental feature which was never
enabled, and was eventually removed. The cow_safety_test tests some
scenarios of canister writes and verifies that the memory is handled
correctly. While it was originally motivated by the COW feature, it does
provide useful test coverage beyond the original feature.

This commit does the following:
- Rename cow_safety_test to memory_safety_test, clarifying what the test
does and remove mentions of removed features.
- Move the test from tests/consensus to tests/message_routing, changing
code ownership in the process.
- Fix the motoko code of the canister, so that it builds with newer dfx.
- Recompile and replace the wasm files that the test uses. Because it's
motoko, this test does not recompile the canister wasm via bazel like we
do with rust canisters.
- Remove one of the two copies of the wasm, and point to the remaining
copy instead from the test.
- Add the test to the hourly and nightly pipeline.