chore(IDX): specify which NNS canisters to use via env vars (#675)
This changes system-tests that depend on the NNS by specifying the path
to the NNS canisters via environment variables. This is both simpler
(see that we removed more lines than we added) and has the advantage
that we can later more easily change which NNS is used for each
system-test without changing Rust code.

The ultimate goal (planned for a later PR) is to let most system-tests
use the mainnet NNS instead of the HEAD NNS. For some tests we want to
have two versions: one that tests against the mainnet NNS and one that
tests against the HEAD NNS. The latter will be made possible if we can
specify the NNS via environment variables instead of hard-coding which
NNS to use in the Rust code.

Why do we want most system-tests to use the mainnet NNS? For tests that
check the guestOS it's more realistic to test against the mainnet NNS
since that's also what would be the case if the guestOS would be
released to mainnet. Futhermore, not using the HEAD NNS means reduced CI
times for PRs touching the NNS itself. So the NNS team will benefit from
this.