test(icrc_ledger): FI-1399: Add an InMemoryLedger for verifying ICRC ledger state (#719)
Add an `InMemoryLedger` that creates an expected ICRC ledger state based
on blocks retrieved from a deployed ICRC ledger and archive(s). Provide
a `verify_ledger_state` helper function that takes a `StateMachine` and
the ledger `CanisterId`, and retrieves the blocks from the ledger and
its archive(s), reconstructs the expected ledger state wrt. account
balances and allowances, and verifies that the balances and allowances
on the deployed ledger matches those of the `InMemoryLedger`.