test(ckerc20): Ensure same processing between EVM-RPC and minter results (#562)
Ensure that the processing of responses from the EVM-RPC canister is
consistent with the processing done by the minter when it does the HTTPs
outcalls itself. This is critical to ensure that the endpoints that
migrated to using the EVM-RPC canister behave as before and also for
backwards-compatibility, in case the migrated endpoints should be back
to doing the HTTPs outcalls themselves.

Remarks regarding the implementation
1. The trait `Reduce` is introduced to unify processing of responses
from the EVM-RPC canister or from the HTTPs outcalls done by the minter,
which use different types.
2. Add one `proptest` per migrated endpoint to test that processing of
responses is always consistent. To ensure that no flakyness was
introduced the new `proptest` have been executed 10_000 times:
    ```
bazel test //rs/ethereum/cketh/minter:lib_tests --test_arg
"evm_rpc_conversion" --runs_per_test 10000
    ```

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>