fix: Update heap delta debit and estimate when handling snapshots (#727)
This PR updates the heap delta debit and estimate when handling canister
snapshots.

In the existing implementation, we would count as heap delta the
snapshot size when taking a snapshot which is incorrect as the heap
delta is supposed to capture the diff of what's going to be persisted at
the next checkpoint (i.e. the changes of the canister that are not yet
persisted). When loading a snapshot, the heap delta debit was not
updated at all which was also a bug.

The PR fixes both endpoints (`take_canister_snapshot` and
`load_canister_snapshot`) by doing the following:
1. Introduce a function that returns the amount of `heap_delta` in the
canister state and in the snapshot respectively.
2. Use the appropriate `heap_delta` value when updating the debit and
estimate for both take_canister_snapshot and load_canister_snapshot.
3. Move updating the `heap_delta_estimate` next to where the
`heap_delta_debit` is updated. Apart from being easier to inspect that
we perform all necessary actions, it also removes the need to propagate
the values from `CanisterManager` back to `ExecutionEnvironment` only to
update this field.