chore: update default dkg-interval-length to use 499 for all subnets (#741)
In the bigger effort of trying to unify configuration between subnets,
we are working towards getting all subnets on the same DKG interval
length of 499. Currently all subnets except the NNS use this
configuration, and we plan to submit a proposal to have the NNS subnet
use this configuration as well. Therefore, it makes sense that we update
our default values for future subnet creations to also use this value of
499.