test(IDX): let most system-tests depend on the mainnet NNS in PRs (#730)
What
====
This duplicates most system-tests which depend on the NNS by:
* Having the original system-test with name `<name>` now depend on the
mainnet NNS instead of the HEAD NNS.
* Having a new system-test with the name `<name>_head_nns` which depends
on the HEAD NNS.
 
`<name>_head_nns` is additionally tagged with `system_test_nightly` such
that it runs nightly and not on PRs.

The above is done for all system-tests which depend on the NNS except
for the ones under `rs/tests/nns` and `rs/tests/nns/sns` because they
test the NNS instead of the replica.

Why
===
The idea is that for most system-tests which test the replica it's more
realistic to test against the mainnet NNS since that version would be
active when the replica would be released. It's also useful to reduce CI
times since changes to the NNS now don't trigger as many system-tests
(although to achieve this fully more work like
https://github.com/dfinity/ic/pull/628 is needed).

However it's still useful to see if the HEAD replica works against the
HEAD NNS (see [this
comment](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/20322#note_1989750018)
in an earlier attempt for an example) which is why this macro introduces
the `<name>_head_nns` variant which only runs daily if not overridden.

How
===

This is follow up work from https://github.com/dfinity/ic/pull/675.

It extends the `system_test(...)` bazel macro with a
`test_driver_target` argument which specifies the name of the target of
the test driver. It defaults to `None` which means that the macro
declares its own test driver as a rust_binary from `<name>.rs`. The
macro then also returns a structure specifying the `test_driver_target`.

This is then used in the new bazel macro `system_test_nns(name, ...)`
which declares two system-tests: one named `<name>` depending on the
mainnet NNS (as specified by
[mainnet-canisters.bzl](https://github.com/dfinity/ic/blob/master/mainnet-canisters.bzl))
and one `head_nns` variant which depends on the head NNS. The latter
then also gets the additional `system_test_nightly` tag which can be
overridden if desired.

Tests
===

Additional testing:

* Schedule Hourly:
https://github.com/dfinity/ic/actions/runs/10231299140
* Schedule Daily: https://github.com/dfinity/ic/actions/runs/10231299567
* Release Testing:
https://github.com/dfinity/ic/actions/runs/10231300143