chore: Remove copy of CreateSubnetPayload in ic-admin (#775)
This is a follow up to https://github.com/dfinity/ic/pull/441. After the
[upgrade of registry
canister](https://dashboard.internetcomputer.org/proposal/131397) to a
version that includes the previous commit that removes the instructions
limit from the subnet record, it's now safe to update `ic-admin` to use
the same struct that does not include the fields anymore.