fix: EXC-1681: Fix a debug assertion in update_socket_timeout (#779)
This change does not affect production because it adjusts a debug mode
assertion.

The assertion was too strict and could fire when running tests in a
corner case
when the main process sends a termination message to the sandbox process
and
manages to close the socket before the sandbox process reads termination
message.

In such cases, an attempt to call `setsockopt` fails with an OS error 9
(bad file descriptor).