test(neurons-fund): Check Neurons' Fund hotkey propagation in SNS lifecycle tests (#705)
This PR adds the newly expected hotkeys from NNS neurons of Neurons'
Fund participants as SNS neuron permissions of SNS neurons created after
a successful swap with Neurons' Fund participation. Additionally, the
code is slightly refactored so that the test can reuse the algorithm for
picking subsets of hotkeys (in cases when there are too many of them on
the NNS side).

< [Previous PR](https://github.com/dfinity/ic/pull/629) |

---------

Co-authored-by: Andre Popovitch <andre@popovit.ch>